Update to SpongeAPI12
Changed the logic of the `/clearlag halt` command. Now it does not disable entity spawning or reduce tickrate to zero. Instead, it stops ticks in the world completely.
The freezing function will also be activated in case of overload. The tick time threshold is set in the configuration. If freezing is activated manually, it cannot be deactivated automatically.